(fix): resupport profiler in incremental mode
