+++ crates/compiler/src/relops.rs
@@ -100,6 +100,8 @@ impl Compiler {
         };
 
         let file = quote! {
+            #![allow(non_camel_case_types)]
+
             //! Dynamic relation handlers for incremental input sessions.
             //!
             //! This module is generated by the FlowLog compiler. It defines a uniform interface
