+++ crates/compiler/src/scaffold.rs
@@ -111,15 +111,14 @@ impl Compiler {
             write_file(&src_dir.join("udf.rs"), content.trim_start())?;
         }
 
-        // Profiler logs — write alongside the final executable so they
-        // survive build-directory cleanup when --save-temps is not set.
+        // Profiler logs — write into a `log/` directory alongside the final
+        // executable so they survive build-directory cleanup when --save-temps
+        // is not set.
         with_profiler_ref(profiler, |profiler| {
             let exe = self.config.executable_path();
-            let ops_path =
-                exe.with_file_name(format!("{}_ops.json", self.config.executable_name()));
-            if let Some(parent) = ops_path.parent() {
-                ensure_dir(parent)?;
-            }
+            let log_dir = exe.with_file_name("log");
+            ensure_dir(&log_dir)?;
+            let ops_path = log_dir.join("ops.json");
             profiler.write_json(&ops_path)
         })?;
