+++ crates/profiler/Cargo.toml
@@ -9,3 +9,5 @@ license.workspace = true
 serde.workspace = true
 serde_json.workspace = true
 tracing.workspace = true
+
+common = { path = "../common" }
